+++ SolastaCommunityExpansion/Patches/DungeonMaker/Editor/DatabaseSelectionModalPatcher.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using HarmonyLib;
 using UnityEngine;
 
-namespace SolastaCommunityExpansion.Patches.DungeonMaker
+namespace SolastaCommunityExpansion.Patches.DungeonMaker.Editor
 {
     [HarmonyPatch(typeof(DatabaseSelectionModal), "SelectMonsterDefinition")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
