+++ SolastaCommunityExpansion/Patches/DungeonMaker/Editor/UserLocationViewPanelPatcher.cs
@@ -2,7 +2,7 @@
 using HarmonyLib;
 using UnityEngine;
 
-namespace SolastaCommunityExpansion.Patches.DungeonMaker
+namespace SolastaCommunityExpansion.Patches.DungeonMaker.Editor
 {
     [HarmonyPatch(typeof(UserLocationViewPanel), "PropOverlap", MethodType.Getter)]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
