+++ SolastaCommunityExpansion/Patches/GameUi/CharacterPanel/RulesetImplementationManagerLocationPatcher.cs
@@ -2,7 +2,7 @@
 using HarmonyLib;
 using UnityEngine;
 
-namespace SolastaCommunityExpansion.Patches.GameUi.Location
+namespace SolastaCommunityExpansion.Patches.GameUi.CharacterPanel
 {
     [HarmonyPatch(typeof(RulesetImplementationManagerLocation), "IsAnyMetamagicOptionAvailable")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
