+++ SolastaCommunityExpansion/Patches/GameUi/Location/CursorPatcher.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using HarmonyLib;
 
-namespace SolastaCommunityExpansion.Patches.GameUi
+namespace SolastaCommunityExpansion.Patches.GameUi.Location
 {
     [HarmonyPatch(typeof(Cursor), "OnClickSecondaryPointer")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
