+++ SolastaCommunityExpansion/Patches/PartySize/GameUi/VictoryModalPatcher.cs
@@ -4,7 +4,7 @@ using HarmonyLib;
 using SolastaCommunityExpansion.Models;
 using UnityEngine;
 
-namespace SolastaCommunityExpansion.Patches.PartySize
+namespace SolastaCommunityExpansion.Patches.PartySize.GameUi
 {
     // this patch scales down the victory modal whenever the party size is bigger than 4
     //
