+++ SolastaCommunityExpansion/Patches/Teleport/FunctorTeleportPatcher.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 using System.Reflection.Emit;
 using HarmonyLib;
 
-namespace SolastaCommunityExpansion.Patches.GameUi.Location
+namespace SolastaCommunityExpansion.Patches.Teleport
 {
     /// <summary>
     /// Currently when a character is teleported off screen the camera doesn't follow.
